Refactor: Relocate UI components to screen packages
This commit moves several composables from the generic `component` package to their more specific `screen` packages, improving code organization and modularity.

- **`feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/resize/`**:
  - `GridItemResizeOverlay.kt` and `WidgetGridItemResizeOverlay.kt` have been moved from `component/resize` to `screen/resize`.
  - The import statements in `ResizeScreen.kt` have been updated to reflect this change.

- **`feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/pager/`**:
  - `Chevron.kt` has been moved from `component/indicator` to `screen/pager`.
  - The `Chevron` composable has been refactored to encapsulate its visibility logic, which was previously managed within `HorizontalPagerScreen.kt`.
