+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/resize/WidgetGridItemResizeOverlay.kt
@@ -15,7 +15,7 @@
  *   limitations under the License.
  *
  */
-package com.eblan.launcher.feature.home.component.resize
+package com.eblan.launcher.feature.home.screen.resize
 
 import android.appwidget.AppWidgetProviderInfo
 import androidx.compose.foundation.background
